+++ PruebaD/Prueba_Proyecto/flask/webapp/frontend/templates/menu_cita.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Menú de Opciones</title>
+    <title>Lista de Citas</title>
     <link rel="stylesheet" href="{{ url_for('static', filename='styles/Smenu_cita.css') }}">
 </head>
 <body>
@@ -12,14 +12,29 @@
         <a href="{{ url_for('home') }}" class="exit-button">Salir</a>
     </div>
 
-    <!-- Menú de opciones -->
+    <!-- Lista de citas -->
     <div class="menu">
-        <h1>Menú de Citas</h1>
-        <a href="crear_cita.html" class="button">Crear Cita</a>
-        <a href="modificar_cita.html" class="button">Modificar Cita</a>
-        <a href="ver_citas.html" class="button">Ver Todas las Citas</a>
-        <a href="{{ url_for('borrar_cita_page') }}" class="button">Borrar Cita</a>
-        <a href="consultar_cita.html" class="button">Consultar Cita</a>
+        <h1>Lista de Citas</h1>
+        
+        <!-- Verificación de si hay citas -->
+        {% if citas %}
+        <div class="cita-list">
+            {% for cita in citas %}
+            <div class="cita-item">
+                <div>
+                    <p><strong>Número de Cita: {{ cita.numeroDeCita }}</strong></p>
+                </div>
+                <div class="cita-buttons">
+                    <a href="{{ url_for('consultar_cita', numeroDeCita=cita.numeroDeCita) }}" class="button">Consultar</a>
+                    <a href="{{ url_for('actualizar_cita', numeroDeCita=cita.numeroDeCita) }}" class="button">Modificar</a>
+                    <a href="{{ url_for('eliminar_cita', numeroDeCita=cita.numeroDeCita) }}" class="button">Borrar</a>
+                </div>
+            </div>
+            {% endfor %}
+        </div>
+        {% else %}
+        <p>No hay citas disponibles en este momento.</p>
+        {% endif %}
     </div>
 </body>
 </html>
