+++ PruebaD/Prueba_Proyecto/flask/webapp/view.py
@@ -3,26 +3,23 @@ from pymongo import MongoClient
 from . import app
 from datetime import datetime
 import os
-from werkzeug.security import generate_password_hash
-from werkzeug.security import check_password_hash
+from werkzeug.security import generate_password_hash, check_password_hash
 from flask import render_template
 
 client = None
 
-#Para trabajar con el contenedor
+# Para trabajar con el contenedor
 @app.before_request
 def initialize_db():
     global client
     mongo_uri = os.environ.get("MONGO_URI")
     client = MongoClient(mongo_uri)
 
-#_______________________________________________________________________________________________________
+# Ruta de inicio
 @app.route('/')
 def home():
-    #return jsonify({"message": "Bienvenido al servicio de generacion de citas medicas "}), 200
     return render_template('inicio.html')
 
-#_______________________________________________________________________________________________________
 # Ruta para cargar el formulario de inicio de sesión
 @app.route('/login', methods=['GET'])
 def login_page():
@@ -47,7 +44,8 @@ def login():
         return jsonify({"message": "Login exitoso"}), 200
     
     return jsonify({"message": "Usuario o contraseña incorrectos"}), 401
-#_________________________________________________________________________________________________________
+
+# Ruta para la página de registro de usuario
 @app.route('/registro', methods=['GET'])
 def crear_usuario_page():
     return render_template('registro.html')
@@ -76,31 +74,43 @@ def crear_usuario():
     usuarios_collection.insert_one(nuevo_usuario)
     return jsonify({"message": "Usuario creado exitosamente"}), 201
 
-#_________________________________________________________________________________________________________
-
-@app.route('/menu_cita')
+# Ruta para el menú de citas
+@app.route('/menu_cita', methods=['GET'])
 def menu_cita():
-    return render_template('menu_cita.html')
+    db = client["proyecto"]
+    citas_collection = db["citas"]
+    citas = []
+    for cita in citas_collection.find():
+        citas.append({
+            "nombre": cita.get("nombre", ""),
+            "fechaCita": cita.get("fechaCita", ""),
+            "hora": cita.get("hora", ""),
+            "descripcion": cita.get("descripcion", ""),
+            "numeroDeCita": cita.get("numeroDeCita", ""),
+            "fechaActualizacionCita": cita.get("fechaActualizacion", "")
+        })
+    return render_template('menu_cita.html', citas=citas)
 
-#_______________________________________________________________________________________________________
 
-# Crear una nueva cita
+
+# Ruta para crear una nueva cita
 @app.route('/crear/cita', methods=['POST'])
 def crear_cita():
     db = client["proyecto"]
     citas_collection = db["citas"]
     data = request.json
-     # Lista de propiedades permitidas
+    
+    # Lista de propiedades permitidas
     propiedades_permitidas = {"nombre", "fecha", "hora", "descripcion"}
     
     # Validación: Verifica que no haya propiedades adicionales
     if not propiedades_permitidas.issuperset(data.keys()):
         return jsonify({"error": "Propiedades no permitidas en el cuerpo de la solicitud"}), 400
 
-     # Agregar fecha de creación
+    # Agregar fecha de creación
     fecha_creacion = datetime.now().strftime("%d/%m/%Y %H:%M:%S")
 
-     # Generar el prefijo de numeroDeCita
+    # Generar el prefijo de numeroDeCita
     fecha_actual = datetime.now().strftime("%y%m%d")
     citas_hoy = citas_collection.count_documents({"numeroDeCita": {"$regex": f"^C{fecha_actual}"}}) + 1
     numero_de_cita = f"C{fecha_actual}{citas_hoy:03}"
@@ -116,7 +126,7 @@ def crear_cita():
     result = citas_collection.insert_one(cita)
     return jsonify({"message": "Cita creada", "numeroDeCita": numero_de_cita}), 201
 
-# Consultar todas las citas
+# Ruta para consultar todas las citas
 @app.route('/ver/citas', methods=['GET'])
 def obtener_citas():
     db = client["proyecto"]
@@ -133,7 +143,19 @@ def obtener_citas():
         })
     return jsonify(citas), 200
 
-# Consultar una cita por ID
+#obtener cita
+@app.route('/consultar/cita/<numeroDeCita>', methods=['GET'])
+def consultar_cita(numeroDeCita):
+    db = client["proyecto"]
+    citas_collection = db["citas"]
+    cita = citas_collection.find_one({"numeroDeCita": numeroDeCita})
+    if cita:
+        return render_template('consultar_cita.html', cita=cita)
+    return jsonify({"message": "Cita no encontrada"}), 404
+
+
+
+# Ruta para consultar una cita por ID
 @app.route('/consultar/cita/<numeroDeCita>', methods=['GET'])
 def obtener_cita(numeroDeCita):
     db = client["proyecto"]
@@ -150,12 +172,13 @@ def obtener_cita(numeroDeCita):
         }), 200
     return jsonify({"message": "Cita no encontrada"}), 404
 
-# Actualizar una cita
+# Ruta para actualizar una cita
 @app.route('/modificar/cita/<numeroDeCita>', methods=['PUT'])
 def actualizar_cita(numeroDeCita):
     db = client["proyecto"]
     citas_collection = db["citas"]
     data = request.json
+    
     # Lista de propiedades permitidas
     propiedades_permitidas = {"nombre", "fecha", "hora", "descripcion"}
     
@@ -163,7 +186,7 @@ def actualizar_cita(numeroDeCita):
     if not propiedades_permitidas.issuperset(data.keys()):
         return jsonify({"error": "Propiedades no permitidas en el cuerpo de la solicitud"}), 400
 
-     # Agregar fecha de creación
+    # Agregar fecha de actualización
     fecha_actualizacion = datetime.now().strftime("%d/%m/%Y %H:%M:%S")
     update_result = citas_collection.update_one(
         {"numeroDeCita": numeroDeCita},
@@ -172,31 +195,27 @@ def actualizar_cita(numeroDeCita):
             "fechaCita": data.get("fecha"),
             "hora": data.get("hora"),
             "descripcion": data.get("descripcion"),
-            "fechaActualizacion":fecha_actualizacion
+            "fechaActualizacion": fecha_actualizacion
         }}
     )
     if update_result.matched_count == 1:
         return jsonify({"message": "Cita actualizada"}), 200
     return jsonify({"message": "Cita no encontrada"}), 404
 
-
-#____________________________________________________________________________________________________
-# Eliminar una cita
+# Ruta para la página de eliminar cita
 @app.route('/borrar_cita', methods=['GET'])
 def borrar_cita_page():
     return render_template('borrar_cita.html')
 
-@app.route('/borrar_cita', methods=['DELETE'])
+# Ruta para eliminar una cita
+@app.route('/borrar_cita', methods=['POST'])
 def eliminar_cita():
-    # Extraer los datos del cuerpo de la solicitud
     data = request.get_json()
     numeroDeCita = data.get("numeroDeCita")
 
-    # Verificar que el número de cita exista
     if not numeroDeCita:
         return jsonify({"message": "Número de cita no proporcionado"}), 400
 
-    # Conectar a la base de datos y eliminar
     db = client["proyecto"]
     citas_collection = db["citas"]
     delete_result = citas_collection.delete_one({"numeroDeCita": numeroDeCita})
@@ -205,8 +224,6 @@ def eliminar_cita():
         return jsonify({"message": "Cita eliminada"}), 200
     return jsonify({"message": "Cita no encontrada. Reintente"}), 404
 
-#___________________________________________________________________________________________________
-
-if __name__== '_main_':
-    app.run(debug=True)
-
+# Iniciar la aplicación
+if __name__ == '__main__':
+    app.run(debug=True, port=8003)
