+++ PruebaD/Prueba_Proyecto/mongo/init-mongo.js
@@ -7,14 +7,16 @@ if(!db.getCollectionNames().includes('citas')){
             "nombre":"Pepito123",
             "fecha":"24/06/2002",
             "hora": "12 pm",
-            "descripcion": "prueba"
+            "descripcion": "prueba",
+            "numeroDeCita": "C241118001"
     });
 
     db.citas.insertOne({
         "nombre":"Pepito123456",
         "fecha":"24/06/2002",
         "hora": "6 pm",
-        "descripcion": "prueba2"
+        "descripcion": "prueba2",
+        "numeroDeCita": "C241118002"
     });
     
 }
