--- docker-compose.yml
@@ -1,47 +1,0 @@
-services:
-  flask:
-    build: ./flask
-    container_name: flask
-    restart: always
-    expose:
-      - 8080
-    depends_on:
-      - mongodb
-    networks:
-      - mi_red
-    environment:
-      - MONGO_URI=mongodb://mongodb:27017
-  frontend:
-    build: ./flask/webapp/frontend/
-    container_name: frontend
-    restart: always
-    expose:
-      - 80
-    networks:
-      - mi_red
-  nginx:
-    build: ./nginx
-    container_name: nginx
-    restart: always
-    ports:
-      - 8003:80
-    networks:
-      - mi_red
-    depends_on:
-      - flask
-      - frontend
-  mongodb:
-    image: mongo:4.2.0
-    container_name: mongodb
-    volumes:
-      - mongo_data:/data/db
-      - ./mongo/init-mongo.js:/docker-entrypoint-initdb.d/init-mongo.js:ro
-    restart: always
-    networks:
-      - mi_red
-    
-
-networks:
-  mi_red:
-volumes:
-  mongo_data:
