--- flask/app.ini
@@ -1,12 +1,0 @@
-[uwsgi]
-wsgi-file = run.py
-module = webapp:app
-callable = app
-socket = :8080
-processes = 4
-threads = 2
-master = true
-chmod socket =660
-vacuum = true
-die-on-term = true
-
