--- flask/run.py
@@ -1,2 +1,0 @@
-from webapp import app
-
