--- flask/webapp/frontend/Dockerfile
@@ -1,11 +1,0 @@
-# Usa una imagen ligera de Nginx
-FROM nginx:alpine
-
-# Copia los archivos estáticos a la carpeta predeterminada de Nginx
-COPY ./static /usr/share/nginx/html/static
-
-# Copia las plantillas HTML al directorio raíz de Nginx
-COPY ./templates /usr/share/nginx/html/templates
-
-# Expone el puerto 80 para acceder al frontend
-EXPOSE 80
