--- flask/webapp/frontend/static/styles/Smenu_cita.css
@@ -1,145 +1,0 @@
-/* General reset */
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-}
-
-/* Main body layout */
-body {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-    background-color: #111;
-    color: #fff;
-    position: relative;
-}
-
-/* Top-left exit button */
-.top-left {
-    position: absolute;
-    top: 20px;
-    left: 20px;
-}
-
-.exit-button {
-    color: #ff5ea5;
-    font-size: 1rem;
-    text-decoration: none;
-    padding: 10px 20px;
-    border-radius: 15px;
-    background-color: #333;
-    border: 1px solid #ff5ea5;
-    transition: background-color 0.3s ease, transform 0.3s ease, color 0.3s ease;
-}
-
-.exit-button:hover {
-    background-color: #e64a8e;
-    color: #fff;
-    transform: translateX(-5px);
-}
-
-/* Menu container */
-.menu {
-    width: 90%;
-    max-width: 900px;
-    padding: 40px;
-    background-color: #1e1e1e;
-    border-radius: 15px;
-    box-shadow: 0 10px 30px rgba(0, 0, 0, 0.6);
-    display: grid;
-    grid-template-columns: 1fr;
-    gap: 20px;
-    align-items: center;
-    text-align: center;
-    animation: fadeIn 1s ease-in-out;
-}
-
-/* Main heading */
-.menu h1 {
-    font-size: 2.5rem;
-    color: #ff5ea5;
-    margin-bottom: 20px;
-    animation: slideIn 1s ease-out;
-}
-
-/* Buttons */
-.menu .button {
-    padding: 15px;
-    font-size: 1.2rem;
-    text-decoration: none;
-    color: #fff;
-    background-color: #ff5ea5;
-    border-radius: 25px;
-    transition: background-color 0.3s ease, transform 0.3s ease, box-shadow 0.3s ease;
-    box-shadow: 0 5px 15px rgba(255, 94, 165, 0.4);
-    text-align: center;
-}
-
-.menu .button:hover {
-    background-color: #e64a8e;
-    transform: scale(1.05);
-}
-
-/* Animations */
-@keyframes fadeIn {
-    from {
-        opacity: 0;
-        transform: translateY(20px);
-    }
-    to {
-        opacity: 1;
-        transform: translateY(0);
-    }
-}
-
-@keyframes slideIn {
-    from {
-        opacity: 0;
-        transform: translateX(-50px);
-    }
-    to {
-        opacity: 1;
-        transform: translateX(0);
-    }
-}
-
-/* Lista de citas y botones */
-.cita-list {
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    margin-top: 20px;
-}
-
-.cita-item {
-    background-color: #2d2d2d;
-    padding: 15px;
-    border-radius: 10px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    box-shadow: 0 5px 15px rgba(0, 0, 0, 0.3);
-}
-
-.cita-buttons {
-    display: flex;
-    gap: 10px;
-}
-
-.cita-buttons button {
-    background-color: #ff5ea5;
-    color: white;
-    padding: 10px 15px;
-    border-radius: 5px;
-    cursor: pointer;
-    border: none;
-    transition: background-color 0.3s ease;
-}
-
-.cita-buttons button:hover {
-    background-color: #e64a8e;
-}
-
