--- flask/webapp/frontend/templates/inicio.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="{{ url_for('static', filename='styles/Sinicio.css') }}">
-    <title>Inicio</title>
-</head>
-<body>
-    <h1>Bienvenido al servicio de generación de citas médicas</h1>
-    <nav>
-        <!-- Redirigir a la ruta de registro definida en Flask -->
-        <a href="{{ url_for('crear_usuario_page') }}">Registro</a>
-
-        <!-- Redirigir a la ruta de login definida en Flask -->
-        <a href="{{ url_for('login_page') }}">Iniciar Sesión</a>
-    </nav>
-</body>
-</html>
