--- flask/webapp/frontend/templates/sesion.html
@@ -1,33 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Login</title>
-    <link rel="stylesheet" href="{{ url_for('static', filename='styles/Ssesion.css') }}">
-</head>
-<body>
-    <!-- Flecha para redirigir al inicio -->
-    <a href="/" class="back-arrow">&#8592; Inicio</a>
-    
-    <div class="login-container">
-        <h2>Iniciar Sesión</h2>
-        <form id="login-form">
-            <label for="username">Usuario:</label>
-            <input type="text" id="username" name="usuario" required>
-            
-            <label for="password">Contraseña:</label>
-            <input type="password" id="password" name="password" required>
-            
-            <button type="submit">Iniciar Sesión</button>
-        </form>
-        
-        <!-- Botón de redirección a registro -->
-        <button onclick="window.location.href='/registro'" class="register-button">Registrarse</button>
-        
-        <p id="response-message"></p>
-    </div>
-    
-    <script src="{{ url_for('static', filename='scripts/SCsesion.js') }}"></script>
-</body>
-</html>
