--- nginx/Dockerfile
@@ -1,6 +1,0 @@
-FROM nginx:latest
-
-RUN rm /etc/nginx/conf.d/default.conf
-
-COPY nginx.conf /etc/nginx/conf.d/
-
