--- nginx/nginx.conf
@@ -1,8 +1,0 @@
-server{
-	listen 80;
-	location / {
-		include uwsgi_params;
-		uwsgi_pass flask:8080;
-	}
-}
-
